requirements.txt
flashcard-generator==0.1.0
werkzeug==2.3.7
google-generativeai==0.3.1
PyPDF2>=3.0.0
reportlab>=3.6.0